About me
My name is Roseline Titilope Oni. I’m a data science graduate with a strong interest in machine learning and how it can be used to solve real-world problems. With a background in Pure and Applied Physics and a diploma in Data Science from Utiva, I enjoy working with tools like Python, SQL, Excel, and Power BI. As I deepen my knowledge in machine learning, I’m excited about the possibilities it brings.